+++ migrations/2020-02-05-234143_create_users/up.sql
@@ -1,0 +1,12 @@
+-- Your SQL goes here
+
+create table users
+(
+	id int auto_increment,
+	name varchar(80) not null,
+	email varchar(128) not null,
+	enabled boolean default false not null,
+	constraint users_pk
+		primary key (id)
+);
+