+++ src/db/models.rs
@@ -1,0 +1,19 @@
+use super::schema::users;
+use serde::{Deserialize, Serialize};
+
+#[derive(Queryable, Serialize, Deserialize, Debug, Clone)]
+pub struct User {
+    pub id: i32,
+    pub name: String,
+    pub email: String,
+    pub enabled: bool,
+}
+
+
+#[derive(Insertable, Serialize, Deserialize, Debug, Clone)]
+#[table_name = "users"]
+pub struct NewUser<'a> {
+    pub name: &'a str,
+    pub email: &'a str,
+    pub enabled: bool,
+}